requirements.txt
pytest-django==4.5.2
Django==4.1.2
djangorestframework==3.12.4
pytz==2022.6
